requirements.txt
streamlit
python-nmap
pandas 